crypto_analysis_project/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dlt==1.8.1
duckdb==1.2.1
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
hexbytes==1.3.0
humanize==4.12.2
idna==3.10
jsonpath-ng==1.7.0
makefun==1.15.6
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.4
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pendulum==3.0.0
pluggy==1.5.0
ply==3.11
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requirements-parser==0.11.0
rich==13.9.4
rich-argparse==1.7.0
semver==3.0.4
setuptools==78.0.2
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
tenacity==9.0.0
time-machine==2.16.0
tomlkit==0.13.2
types-setuptools==76.0.0.20250313
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
